refactor: extract constants, utils
