+++ src/js/constants/constants.js
@@ -1,0 +1,11 @@
+export const ERROR_MESSAGE = {
+  NEGATIVE_INPUT_ERROR: 'is money negative',
+  NOT_INTEGER_INPUT_ERROR: 'is money not a integer',
+};
+
+export const CONDITIONS = {
+  LOTTO_SIZE: 7,
+  LOTTO_NUM_MIN: 1,
+  LOTTO_NUM_MAX: 45,
+  LOTTO_PRICE: 1000,
+};
