+++ src/js/model/Lotto.js
@@ -1,0 +1,16 @@
+import { getValues } from '../utils.js';
+import { CONDITIONS } from '../constants/constants.js';
+
+export class Lotto {
+  constructor() {
+    this.numbers = [];
+  }
+
+  makeNumbers() {
+    const temp = new Set();
+    while (temp.size < CONDITIONS.LOTTO_SIZE) {
+      temp.add(getValues.randomInt(CONDITIONS.LOTTO_NUM_MIN, CONDITIONS.LOTTO_NUM_MAX));
+    }
+    this.numbers = Array.from(temp);
+  }
+}
